+++ R/get_sig_similarity.R
@@ -5,14 +5,20 @@
 #' Two COSMIC databases are used for comparisons - "legacy" which includes 30 signaures,
 #' and "SBS" - which includes updated/refined 65 signatures. This function is modified
 #' from `compareSignatures()` in **maftools** package.
+#' **NOTE**: all reference signatures are generated from gold standard tool:
+#' SigProfiler.
 #'
-#' @param Signature a `Signature` object or a component-by-signature matrix (sum of each column is 1)
-#' or a normalized component-by-sample matirx (sum of each column is 1).
+#' @param Signature a `Signature` object or a component-by-signature matrix/`data.frame`
+#' (sum of each column is 1) or a normalized component-by-sample matrix/`data.frame`
+#' (sum of each column is 1).
 #' More please see examples.
 #' @param Ref default is `NULL`, can be a same object as `Signature`.
-#' @param sig_db can be 'legacy' (for COSMIC v2 'SBS'),
+#' @param sig_db default 'legacy', it can be 'legacy' (for COSMIC v2 'SBS'),
 #' 'SBS', 'DBS', 'ID' and 'TSB' (for SBS transcriptional strand bias signatures).
-#' Default 'legacy'.
+#' For more specific details, it can also be 'SBS_hg19', 'SBS_hg38',
+#' 'SBS_mm9', 'SBS_mm10', 'DBS_hg19', 'DBS_hg38', 'DBS_mm9', 'DBS_mm10'.
+#' But the signature profile for different genome builds are basically same.
+#'
 #' @param db_type only used when `sig_db` is enabled.
 #' "" for keeping default, "human-exome" for transforming to exome frequency of component,
 #' and "human-genome" for transforming to whole genome frequency of component.
@@ -21,6 +27,7 @@
 #' @param normalize one of "row" and "feature". "row" is typically used
 #' for common mutational signatures. "feature" is designed by me to use when input
 #' are copy number signatures.
+#' @param set_order if `TRUE`, order the return similarity matrix.
 #' @param pattern_to_rm patterns for removing some features/components in similarity
 #' calculation. A vector of component name is also accepted.
 #' The remove operation will be done after normalization. Default is `NULL`.
@@ -45,32 +52,43 @@
 #' s3 <- get_sig_similarity(sig2, sig_db = "SBS")
 #' s3
 #'
+#' # Set order for result similarity matrix
+#' s4 <- get_sig_similarity(sig2, sig_db = "SBS", set_order = TRUE)
+#' s4
+#'
 #' ## Remove some components
 #' ## in similarity calculation
-#' s4 <- get_sig_similarity(sig2,
+#' s5 <- get_sig_similarity(sig2,
 #'   Ref = sig2,
 #'   pattern_to_rm = c("T[T>G]C", "T[T>G]G", "T[T>G]T")
 #' )
-#' s4
+#' s5
 #'
 #' ## Same to DBS and ID signatures
+#' x1 <- get_sig_db("DBS_hg19")
+#' x2 <- get_sig_db("DBS_hg38")
+#' s6 <- get_sig_similarity(x1$db, x2$db)
+#' s6
 #' @testexamples
 #' expect_equal(length(s1), 4L)
 #' expect_equal(length(s2), 4L)
 #' expect_equal(length(s3), 4L)
 #' expect_equal(length(s4), 4L)
+#' expect_equal(length(s5), 4L)
+#' expect_is(s6, "list")
 get_sig_similarity <- function(Signature, Ref = NULL,
                                sig_db = "legacy",
                                db_type = c("", "human-exome", "human-genome"),
                                method = "cosine",
                                normalize = c("row", "feature"),
                                feature_setting = sigminer::CN.features,
+                               set_order = TRUE,
                                pattern_to_rm = NULL,
                                verbose = TRUE) {
   if (inherits(Signature, "Signature")) {
     w <- Signature$Signature.norm
-  } else if (is.matrix(Signature)) {
-    w <- Signature
+  } else if (is.matrix(Signature) | is.data.frame(Signature)) {
+    w <- as.matrix(Signature)
     if (!all(round(colSums(w)) == 1)) {
       stop("If Signature is a matrix, sum of each column is 1!", call. = FALSE)
     }
@@ -93,7 +111,10 @@ get_sig_similarity <- function(Signature, Ref = NULL,
     }
   }
 
-  sig_db <- match.arg(arg = sig_db, choices = c("legacy", "SBS", "DBS", "ID", "TSB"))
+  sig_db <- match.arg(arg = sig_db,
+                      choices = c("legacy", "SBS", "DBS", "ID", "TSB",
+                                  "SBS_hg19", "SBS_hg38", "SBS_mm9", "SBS_mm10",
+                                  "DBS_hg19", "DBS_hg38", "DBS_mm9", "DBS_mm10"))
   method <- match.arg(arg = method, choices = c("cosine"))
 
   if (is.null(Ref)) {
@@ -113,8 +134,8 @@ get_sig_similarity <- function(Signature, Ref = NULL,
   } else {
     if (inherits(Ref, "Signature")) {
       sigs <- Ref$Signature.norm
-    } else if (is.matrix(Ref)) {
-      sigs <- Ref
+    } else if (is.matrix(Ref) | is.data.frame(Ref)) {
+      sigs <- as.matrix(Ref)
       if (!all(round(colSums(sigs)) == 1)) {
         stop("If Ref is a matrix, sum of each column is 1!", call. = FALSE)
       }
@@ -197,25 +218,32 @@ get_sig_similarity <- function(Signature, Ref = NULL,
   if (verbose) {
     if (is.null(Ref)) {
       message("-Comparing against COSMIC signatures")
+      message("------------------------------------")
     } else {
-      message("-Comparing against Custom signatures")
+      message("-Comparing against Custom signatures/profiles")
+      message("---------------------------------------------")
     }
-    message("------------------------------------")
+
+    all_matches <- c()
     for (i in 1:nrow(corMat)) {
+      .to_match <- rownames(corMat)[i]
+      .be_match <- names(which(corMat[i, ] == max(corMat[i, ])))
+      all_matches <- c(all_matches, .be_match)
+
       if (is.null(Ref)) {
-        ae <- aetiology[names(which(corMat[i, ] == max(corMat[i, ]))), ]
+        ae <- aetiology[.be_match, ]
         ae <- paste0(
           "Aetiology: ", ae, " [similarity: ",
           max(corMat[i, ]), "]"
         )
-        message("--Found ", rownames(corMat)[i], " most similar to ",
-          names(which(corMat[i, ] == max(corMat[i, ]))),
+        message("--Found ", .to_match,
+                " most similar to ", .be_match,
           sep = ""
         )
         message(paste0("   ", ae))
       } else {
-        message("--Found ", rownames(corMat)[i], " most similar to ",
-          names(which(corMat[i, ] == max(corMat[i, ]))),
+        message("--Found ", .to_match,
+                " most similar to ", .be_match,
           paste0(" [similarity: ", max(corMat[i, ]), "]"),
           sep = ""
         )
@@ -223,6 +251,11 @@ get_sig_similarity <- function(Signature, Ref = NULL,
     }
     message("------------------------------------")
     message("Return result invisiblely.")
+  }
+
+  if (set_order) {
+    all_matches <- unique(all_matches)
+    corMat <- corMat[, c(all_matches, setdiff(colnames(corMat), all_matches)), drop = FALSE]
   }
 
   res <- list(
@@ -241,7 +274,8 @@ get_sig_similarity <- function(Signature, Ref = NULL,
 
 #' Obtain Reference Signatures
 #'
-#' The signatures and their aetiologies mainly obtained from COSMIC database and cleaned before saving into
+#' The signatures and their aetiologies mainly obtained from COSMIC database
+#' (SigProfiler results) and cleaned before saving into
 #' **sigminer** package.
 #'
 #' @inheritParams get_sig_similarity
@@ -253,11 +287,14 @@ get_sig_similarity <- function(Signature, Ref = NULL,
 #' @examples
 #' s1 <- get_sig_db()
 #' s2 <- get_sig_db("DBS")
+#' s3 <- get_sig_db("DBS_mm10")
 #' s1
 #' s2
+#' s3
 #' @testexamples
 #' expect_is(s1, "list")
 #' expect_is(s2, "list")
+#' expect_is(s3, "list")
 get_sig_db <- function(sig_db = "legacy") {
   db_file <- switch(
     sig_db,
@@ -275,9 +312,26 @@ get_sig_db <- function(sig_db = "legacy") {
     ),
     TSB = system.file("extdata", "TSB_signatures.rds",
       package = "sigminer", mustWork = TRUE
-    )
+    ),
+    {
+      if (startsWith(sig_db, "SBS")) {
+        system.file("extdata", "SBS_signatures_list.rds",
+                    package = "sigminer", mustWork = TRUE
+        )
+      } else if (startsWith(sig_db, "DBS")) {
+        system.file("extdata", "DBS_signatures_list.rds",
+                    package = "sigminer", mustWork = TRUE
+        )
+      } else {
+        stop("Bad input for option 'sig_db'!")
+      }
+    }
   )
   sigs_db <- readRDS(file = db_file)
+  if (!"db" %in% names(sigs_db)) {
+    sigs_db$db <- sigs_db[[sub(".*_", "", sig_db)]]
+    sigs_db <- sigs_db[c("db", "aetiology", "date")]
+  }
   ## Make sure column-sum is 1, i.e. normalized
   sigs_db$db <- apply(sigs_db$db, 2, function(x) x / sum(x))
   sigs_db